refactor: split code-detail into multiple comp
