+++ src/components/discountCode/detail/summary.css
@@ -1,0 +1,27 @@
+.long-content {
+  width: 100%;
+  position: relative;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  float: left;
+  padding: 0 5px;
+  text-overflow: ellipsis;
+
+  &:after {
+    content: '\02026';
+    position: absolute;
+    top: 3.15em;
+    right: 1px;
+    width: 4em;
+    background: linear-gradient(
+      to right,
+      rgba(255, 255, 255, 0),
+      #fff 50%,
+      #fff
+    );
+    text-align: right;
+    vertical-align: bottom;
+  }
+}
